requirements.txt
amqp==5.3.1
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
fastapi==0.120.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
kombu==5.5.4
openai==2.6.0
packaging==25.0
prompt_toolkit==3.0.52
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==7.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
